+++ app/middlewareProtected/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function page() {
+  return (
+    <div>hi</div>
+  )
+}
+
+export default page
