+++ middleware.ts
@@ -1,0 +1,8 @@
+export { default } from "next-auth/middleware";
+
+export const config = {
+  // matched paths are route protected
+  matcher: [
+    '/middlewareProtected',
+  ]
+};
